docker_app/requirements.txt
Flask
mysql-connector
uWSGI==2.0.17.1
